requirements.txt
mcp[cli]==1.9.0
httpx==0.27.0
pydantic==2.10.5
uvicorn==0.34.0
starlette==0.41.3
python-dotenv==1.0.0
